Rename field file to index
